requirements.txt
torch>=2.1.0
transformers>=4.44.0
datasets>=2.19.0
accelerate>=0.33.0
peft>=0.12.0
trl>=0.9.6
tqdm>=4.66.0
pyyaml>=6.0.1
numpy>=1.26.0
